Merged PR 60435: tyger cli authentication
Adding interactive, device code, and service principal + certificate authentication to the tyger cli. The server now has an endpoint publishing the authority and audience at `/v1/metadata`, which the cli hits in order to know how to log in.
We cache the refresh token and cert path in the existing `.tyger` file.

Also adding AAD apps to our eminencemsr AAD tenant to be able to user and test this.

There's also a simple script that exercises the various scenarios. Video below.

![Screen Recording 2022-01-27 at 3.46.57 PM.gif](https://dev.azure.com/msresearch/b54927b4-7877-4e5f-b8fc-226850558453/_apis/git/repositories/9652e2d6-5c0f-4ada-ae48-53e308ca25c2/pullRequests/60435/attachments/Screen%20Recording%202022-01-27%20at%203.46.57%20PM.gif)