Merged PR 60505: Use CLI app id when signing in.
It turns out that the tyger cli app was not able to use its refresh token after the initial access token expires. The reason turns out to be because the I was using the cli identifier URI instead of the app ID (GUID). The former works for getting the token, but the refresh token is unusable.

So I hardcode the app ID in the tyger cli app. I made the app multi-tenant, so that this ID remains constant across tenants (in the public AAD instance anyway).