Merged PR 60660: Reimplement tyger server in C#
Rewriting the tyger server in C#. The main reason for doing so is to be able to validate JWT using approved libraries.

Keeping the cli in go and maintaining the same E2E test suite.

There are changes to the configuration interface and the OpenApi spec is generated from the source, but otherwise the contract is meant to be the same.