Merged PR 60703: Enforcing C# style
Adding an `.editorconfig` file with rules for C# formatting and style.
Most of the rules can be fixed automatically with editor actions (ctrl + .) or by running `make format`.
Style violations will not prevent the code from building but we will validate that they are all taken care of in the pipeline.

Adopting most of the rules from the asp.net roslyn teams. One exception is that I don't think we need to require `var` everywhere, since in some cases specifying the type can make code easier to review.

Changes to the code as a result of the rules:
- [Changed directory casing and namespaces](https://dev.azure.com/msresearch/compimag/_git/tyger/commit/f129dcf173d98acc93d80e1063d79a58ccc72c23?refName=refs/heads/jostairs/csharp-code-style)
- [Fix a real bug that the analyzers exposed](https://dev.azure.com/msresearch/compimag/_git/tyger/commit/fe40cb93525b6c595c9dd78047c4283bb383c210?refName=refs/heads/jostairs/csharp-code-style)
- [Automatic fixes from running `make format`](https://dev.azure.com/msresearch/compimag/_git/tyger/commit/f13225008b9135add8f8f1b31cad5104dbc8ef6b?refName=refs/heads/jostairs/csharp-code-style)
- [A couple of fixes that needed to be done in the editor](https://dev.azure.com/msresearch/compimag/_git/tyger/commit/07f5de67cde67112fa94f4fa301c0e46160192ef?refName=refs/heads/jostairs/csharp-code-style) (but there was a tooltip).