Merged PR 62221: Move to AKS
Changing inner loop development environment to target AKS clusters and Azure Storage instead of k3s and Azurite.
Also updating our environment model and specification mechanism. We are now specifying environments in [CUE](https://cuelang.org/), a configuration language with the ability to specify, constrain, validate, and generate.
Ephemeral environments (like for PRs, local development, etc) are generated with the following script:

``` bash
 $ scripts/get-context-environment-config.sh
```

Which, for me, outputs:

![image.png](https://dev.azure.com/msresearch/b54927b4-7877-4e5f-b8fc-226850558453/_apis/git/repositories/9652e2d6-5c0f-4ada-ae48-53e308ca25c2/pullRequests/62221/attachments/image.png)

(Apologies for using a screenshot, apparently a PR description cannot exceed 4000 characters)

A few things to call out:
- The name of the environment comes from an environment variable or your git email alias
- An environment is made up of multiple clusters, one of which is a primary cluster, which hosts the tyger control place. The other clusters are purely for runs. I have not added support for actually calling into other clusters, so any non-primary clusters are effectively ignored.
- An environment can have multiple organizations (think tenants), which share the same clusters but use separate namespaces, storage accounts, domain name, etc. Our dev environments have a "Lamna" organization, from [Approved fictitious names & guidelines](https://microsoft.sharepoint.com/sites/CELAWeb-Copyrights-Trademarks-And-Patents/SitePages/trademarks-fictitious-names.aspx).
- An organization can have multiple storage accounts for buffers, but I have not yet implemented support for this yet. Only the first one is used.
- For dev environments, only the name needs to be provided, all other properties are defaulted, but can be explicitly provided. Customer environments would have less defaulting but would have the same base constraints.
- Customers could specify their environments in CUE, JSON, or YAML, and all would be validated against the same constraints. (We only support CUE at the moment).

This is all a little overkill right now, but adopting a slightly stronger model allows us to have fewer hardcoded values in our scripts and should make it easier to ship these scripts to partners without as much rework.

The `clusters` specification above is now deployed as a configmap and the tyger server now allows targeting clusters and/or nodepools for runs.

We are NOT ready to retire the eminence repo yet, as I have not yet ported over the end-to-end recon tests and dependency management. That will be coming next.