Merged PR 62691: Add logs endpoint for runs
Adding the ability to retrieve the logs for a run. Logs will be fetched from the pod while it is still alive, otherwise a storage account blob. Before deleting the pod, we copy the logs to the storage account.

We support the flags `--timestamps`, `--since`, `--tail-lines`, `--follow` and `--previous`.

When retrieving archived logs, we're taking care not to have to buffer large amounts of data in order to evaluate `--tail-lines` or very long lines.

When following logs of a container that has not started yet, we start a watch and wait for the container to start. This is different from the k8s API, which returns a 400 status code.

Areas of weakness:
- The "watch until started" functionality is a little flakey.
- ~~There is a potential for race conditions when retrieving a logs from a pod, and in another thread (or process) we delete the pod. We could introduce a locking protocol.~~
**UPDATE**: we now store the time of archival in the database. We'll read from the DB to know where to read the logs. We also let 30s elapse after archiving the logs before deleting the pod, to drain any active requests.
- Timestamps in logs have nanosecond precision, but .NET datetimes only have a resolution of 100ns. We could use NodaTime instead, but the parsing is less efficient.