Merged PR 63044: Pinning AKS preview version
Pinning AKS preview version. There is this [issue](https://github.com/Azure/azure-cli/issues/22032) which affects `az` in general but also the aks-preview extension. Pramod Valavala has confirmed it is a bug and they are working on a fix. For now we will pin versions even of the preview extension.