Merged PR 63819: Use current tag to look up dependencies
Proposing that we adopt the `current` tag to indicate the latest version of our dependency images that were pushed by the CI pipeline.

The existing approach no longer works because the mrd-storage-server images are not tagged with the git commit hash, but instead with the GitHub release name (`vX.Y.Z`).

When CI for wabakimi runs, it will tag the image with the git hash as it does today, but it would also tag the image with `current`. Then in `check-for-dependency-updates.sh`, we look up the `current` tag for each of our dependencies, but update `dependencies.json` with the git hash or version specifier that the image is also tagged with.