Merged PR 64036: Scale nodes during pipeline runs
So that end-to-end tests run don't time out waiting for the cluster autoscaler, we will pre-scale the node pools before tests run. To do this, we will set the minumum node count to 1 and wait for the node count to reach the minimum before starting the tests.

To do this, we create a sub-directory within our CUE configuration directory, where we override the minimum node count.

At the end of the pipeline, we scale back down to zero, unless we are about to delete the environment.