Merged PR 64261: Fix .tyger file race condition
Fixing a race condtion discovered when testing !63464, which starts using the tyger cli concurrently. When a process is reading from the `.tyger` file while another is writing to it, you can occationally get 0-byte reads without any error. (One discussion of this is here: https://stackoverflow.com/questions/32319031/zero-length-read-from-file).

To address this, we could either use a file lock or write to a temp file and rename, which is atomic. I chose a file lock, since this file has sensitive data and it seems like having this data in random temp files is a bad idea. (The process could crash, etc.).