Merged PR 63464: Support basic gadgetron test cases
These changes bring in the cases we wish to support in gadgetron.

These changes include a script to download the tests data dependencies, and convert the tests to a simpler format which can be leveraged in Eminence's pytests.

The script:
  - Identifies all test cases and parses them into a simple data structure representing the config
  - Identifies new test cases and downloads data for them / generates config  files to execute them
  - Downloads noise dependency data, reconstruction data, and reference images identified in the configs
  - Validates downloads against a known hash
  - Converts noise/recon dependencies to .h5 format using siemens_to_ismrmrd
  - Generates 'case files' (case.yml) for each test case
  - Generates run files (for use in eminence) for each test case
 - Generates `testdata.json` which lists the supported tests for consumption in pytest.

Also included is a test which finds converted cases and parameterizes a test to run them all.