Merged PR 65276: Fix codespec list tests; warn on truncated list; use go generics
The codespec list tests were failing in CI because they limit the output to 1000 items and we have more codespecs in than that the CI database. I changed the test to use a prefix so that we don't have to page through all codespecs.

Also, we now emit a yellow warning to stderr when listing runs or codespecs and the output is truncated and the `--limit` parameter was not specified.

Finally, I refactored the common client paging code to use Go generics.