Merged PR 71123: tyger run watch
Adding a watch capability to the get runs operation, which writes out the json in a single line whenever the status changes (until a terminal state is reached). This capability is used from the integration tests, from `tyger run exec`, and from a new `tyger run watch` command, which only outputs the "system" fields to keep the output brief.

In the example below, you will see run state changes written out to the exec logs on the top, and on the bottom, the same run watched with `tyger run watch`

![Screen Recording 2023-03-10 at 5.41.44 PM.gif](https://dev.azure.com/msresearch/b54927b4-7877-4e5f-b8fc-226850558453/_apis/git/repositories/9652e2d6-5c0f-4ada-ae48-53e308ca25c2/pullRequests/71123/attachments/Screen%20Recording%202023-03-10%20at%205.41.44%20PM.gif)