Merged PR 71223: Add `--full-resource` option to `tyger run watch`
Since using the new `tyger run watch` capability in the integration tests, I have found that the "TestGpuResourceRequirement" test fails occasionally. At the end of the watch, we were calling `tyger run show`, but it would seem like this was returning an earlier version of the resource than what `watch` returned.  This earlier version, based on an earlier version of the Pod resource, did not have the `nodePool` field set, because the Pod's `node` field was not set.
Adding --full-resource option to `run watch` as a workaround, so we do not need the `show` call after the watch.