Merged PR 71244: Include sidecar logs in `tyger run logs`
We were only including the logs of the main recon container in `tyger run logs`. With this change, we include the sidecar logs as well, merging them in as we already do for job/worker pod logs.

Example (color NOT part of this PR, added in manually to bring attention to the change):

![image.png](https://dev.azure.com/msresearch/b54927b4-7877-4e5f-b8fc-226850558453/_apis/git/repositories/9652e2d6-5c0f-4ada-ae48-53e308ca25c2/pullRequests/71244/attachments/image.png)

Also correcting the handling of the default value of `--replicas` parameter in `tyger run exec/create`